+++ app/passport.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace App;
+
+use Illuminate\Database\Eloquent\Model;
+
+class passport extends Model
+{
+    public function user()
+    {
+    	return $this->belongsTo('App\user');
+    }
+}
