+++ resources/views/test.blade.php
@@ -1,0 +1,8 @@
+@extends('layout.app')
+
+@section('title', '1-1-rel')
+
+@section('body')
+    Welcome {{ $user->name }} <br>
+    Your passport #: {{ $usernum->number }}
+@endsection
