+++ routes/web.php
@@ -11,6 +11,6 @@
 |
 */
 
-Route::get('/', function () {
-    return view('welcome');
-});
+
+Route::get('/', 'HomeController@home');
+Route::get('/test', 'HomeController@test');
